changed some int declarations to size_t to avoid possible segfaults, solved a segfault as a result of a space in quotations (as second argument) and also pushing CMakeLists.txt (setup for debugging in CLion)

+++ src/sort/ps_radix_sort.c
@@ -14,9 +14,9 @@
 
 void	ps_radix_sort(t_node **stack_a, t_node **stack_b)
 {
-	int	count;
-	int	i;
-	int	j;
+	size_t	count;
+	size_t	i;
+	size_t	j;
 
 	count = list_size(*stack_a);
 	j = 0;
